test(biblecore): split workspace window tests
Why:
- Continue the test architecture migration by moving workspace, window, bookmark-label, and text-display inheritance coverage out of the app-host test bundle.
- Keep package-owned behavior in fast package targets while leaving repository source manager coverage in the app-host bundle until its owning slice is migrated.

What Changed:
- Added BibleCore package coverage for workspace settings, bookmark label filtering, workspace graph lifecycle, window manager routing/readiness, workspace selection, and global text-display propagation.
- Added BibleUI package coverage for workspace text-display propagation helper behavior.
- Removed the migrated tests from AndBibleTests+WorkspaceAndRepository while retaining repository source manager tests there.
- Updated the migration plan with the split and remaining repository source manager follow-up.

Validation:
- python3 scripts/check_repo_standards.py docblocks --all-files
- xcodebuild test -project AndBible.xcodeproj -scheme BibleCoreTests -destination platform=iOS Simulator,name=iPhone 17 -derivedDataPath /tmp/andbible-workspace-core-biblecore-docblocks -only-testing:BibleCoreTests/WorkspaceWindowStoreTests CODE_SIGNING_ALLOWED=NO
- xcrun swift package describe --type json
- git diff --cached --check
- npx gitnexus analyze
- gitnexus detect_changes scope staged: low risk, no affected execution flows
- Prior in-branch validation: BibleCoreTests full suite passed; BibleUITests full suite passed on rerun after one transient simulator/settings test collection blip; AndBible build-for-testing passed; scripts unittest suite passed; pbxproj lint passed.

Impact:
- Reduces app-host test bundle scope without changing production behavior.
- Moves durable workspace/window contracts toward the faster package-test lane and documents the remaining app-hosted repository source manager slice.
