+++ docs/test-architecture-migration-plan.md
@@ -148,6 +148,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+RemoteSyncLifecycle` has moved to `BibleCoreTests` because it exercises category synchronization lifecycle, WebDAV adapter factory selection, reading-plan patch orchestration, and background-refresh coordination through BibleCore services without UI or app bootstrap.
 - `AndBibleTests+RemoteSyncBookmarks` has moved to `BibleCoreTests` because it exercises Android-compatible bookmark snapshot restore, sparse patch replay/upload, initial backup handling, and bookmark category synchronization through BibleCore services without UI or app bootstrap.
 - `AndBibleTests+ExternalDocumentImport` has been split by owning module: document import routing, Android toast feedback, provider filename normalization, and installer selection moved to `BibleUITests`, while Android-style TTF font repository filesystem contracts moved to `SwordKitTests`.
+- `AndBibleTests+WorkspaceAndRepository` has been split by owning module: workspace settings/store/window-manager/selection, bookmark-label filtering, and global text-display propagation moved to `BibleCoreTests`; the UI-owned workspace text-options dirty-field propagation helper moved to `BibleUITests`; repository-source-manager coverage remains app-hosted for a later SwordKit-oriented slice.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; `testContentViewDoesNotContainLegacyRootSidebarShell` becomes a repo-standards guardrail or stays app-hosted (decide explicitly - do not move to BibleUITests); the other ~78 funcs go to BibleUITests.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
